Add files via upload
#Imagine that you work as Data Mining and Machine Learning independent consultant, providing
scientific advisory and consulting services to companies seeking to apply data analytics to their
business activities.The manager of a social media company contacts you to investigate the feasibility of using
machine learning to better identify fake news and misleading information. Over the past year,
the company has had a run of bad press surrounding their inability to address the problem. The
manager is tasked with finding methods to address the problem and ultimately improve the image of
the company. False and misleading information posted on their platform comes from many different
sources including, but not limited to, newspapers, TV stations and individual users. They want
your help to address the problem of identifying accounts that are likely posting false information.
The manager has access to historical data for all accounts, and a list of accounts known to be
troublesome. They wish for you to find a solution that identifies other problematic accounts.

Tasks to be Performed:
1: To identify machine learning techniques appropriate for
a particular practical problem.
2:Undertake a comparative evaluation of several machine
learning procedures when applied to the specific problem.In my study I have used Decision Tree with pruning,Logistic Regression,Support Vector Machines, Naive Bayes Algorithm.
3:Required to produce class predictions of the records in the test set using one approach of your choice among those tested in the comparative study (for example the one achieving the best performance). These data must not be used other than to test the algorithm
trained on the training data.